requirements.txt
# SeenSlide Desktop App Requirements
# Python 3.10+ required

# GUI Framework
customtkinter>=5.2.0

# Local Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Models
pydantic>=2.0.0

# Screen Capture
mss>=9.0.0

# Image Processing & Deduplication
Pillow>=10.0.0
imagehash>=4.3.1

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Database (local SQLite is built-in)
# No additional database dependencies needed

# Utilities
python-dateutil>=2.8.2
